pyproject.toml
[build-system]
requires = [ "setuptools>=61" ]

[project]
name = "textbite"
version = "0.1.0"
description = "Toolkit for extracting logical chunks from complex document pages."
readme = "README.md"
requires-python = ">=3.7"
license = {file = "LICENSE"}
keywords = ["Layout analysis"]

authors = [
  {email = "ibenes@fit.vutbr.cz", name = "Karel Beneš"},
]

classifiers=[
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
]

dependencies = [
    "numpy",
    "shapely",
    "pero-ocr",
    "ultralytics",
    "transformers",
    "torch_geometric",
]

[tool.setuptools]
include-package-data = false

[tool.setuptools.package-dir]
"textbite" = "textbite"

[project.scripts]
textbite = "textbite.infer:main"
